requirements.txt
# in any of your work requiring any functions, kindly add in the package name + version here
# this allows us to quickly install them into a virtual environment where we can just type
# pip install -r requirements.txt and get everything we need for this project.

# the following are just examples and not for this pii project:
requests==2.20.0
jsonschema==2.6.0
simplejson==3.16.0
emoji==0.5.1
python-telegram-bot==11.1.0


spacy == ?
pdfminer.six == ?
nltk == ?